moved package_example to package_test.go
